Refactor `<ContactSupport />` component (#1806)
* remove contact card comp

* refactor contact card

* streamlined styles
